popUpForm close to final design. It pops when clicking on Add flight button
