+++ frontend-app-skynuc/src/components/popUpForm/popUpForm.css
@@ -1,0 +1,136 @@
+.popup-box {
+    position: fixed;
+    background: #00000050;
+    width: 100%;
+    height: 100vh;
+    top: 0;
+    left: 0;
+    z-index: 30;
+}
+
+.box {
+    position: relative;
+    /* margin: 0 auto; */
+    height: auto;
+    /* max-height: 170vh; */
+    margin-top: calc(100vh - 61vh - 40px);
+    background: #fff;
+    overflow: auto;
+    /* background: red; */
+
+    box-shadow: 0 4px 30px rgba(15, 22, 56, 0.35);
+    max-width: 60%;
+    margin-left: auto;
+    margin-right: auto;
+    border-radius: 10px;
+    padding: 25px;
+    /* transform: scale(0.8); */
+    /* position: relative; */
+    z-index: 3;
+    transition: all 0.3s;
+    transition-delay: 0.1s;
+
+    /* display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;  */
+    /* FIXME */
+}
+
+.close-icon {
+    content: 'x';
+    cursor: pointer;
+    position: fixed;
+    right: calc(20% - 12px);
+    top: calc(100vh - 62vh - 42px);
+    background: #ededed;
+    width: 27px;
+    height: 27px;
+    border-radius: 50%;
+    line-height: 20px;
+    text-align: center;
+    border: 1px solid #999;
+    font-size: 21px;
+}
+
+.top-bar {
+    height: 55px;
+    width: 100%;
+    position: absolute;
+    top: 0px;
+    left: 0px;
+    background: linear-gradient(45deg, #217008, #0ead31);
+    color: #fff;
+    font-weight: 600;
+    font-size: 17px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.profile-card-form__container {
+    /* background-color: pink; */
+    height: auto;
+    width: auto;
+    position: relative;
+    /* top: 3em; */
+
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    justify-content: space-around;
+    align-items: center;
+    margin: 65px auto 20px auto;
+}
+
+.profile-card-form__container input,
+label,
+select {
+    margin: 15px 5px;
+    padding: 3px 3px;
+    font-weight: 500;
+    font-size: 17px;
+    resize: none;
+    color: #324e63;
+    transition: all 0.3s;
+}
+
+.profile-card-form__container input,
+select {
+    border: 2px solid #dcdcdc;
+    border-radius: 10px;
+    outline: none;
+}
+
+.profile-card-form__container label {
+    /* min-width: 400px; */
+}
+
+.profile-card-form__container input:focus,
+select:focus {
+    outline: none;
+    border-color: #0ead31;
+}
+
+.profile-card-form-dpt__container,
+.profile-card-form-arr__container {
+    flex-basis: 50%;
+}
+
+.profile-card-form-price__container {
+    flex-basis: 100%;
+    /* background: brown; */
+    display: flex;
+    justify-content: center;
+}
+
+.profile-card-form__bottom {
+    display: flex;
+    justify-content: center;
+    /* margin-top: 30px; */
+}
+@media screen and (max-width: 576px) {
+    .profile-card-form__bottom {
+        flex-wrap: wrap;
+    }
+}
