--- frontend-app-skynuc/src/components/popUpForm_TEST/popUpForm.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import './popUpForm.css';
-
-const PopUpForm = (props) => {
-    return (
-        <div className="popup-box">
-            <div className="box">
-                <span className="close-icon" onClick={props.handleClose}>
-                    x
-                </span>
-                {props.content}
-            </div>
-        </div>
-    );
-};
-
-export default PopUpForm;
