+++ frontend-app-skynuc/src/components/userProfileCard/userProfileCard.css
@@ -84,7 +84,7 @@ a:hover {
     border-radius: 50%;
     overflow: hidden;
     position: relative;
-    z-index: 4;
+    z-index: 40;
     box-shadow: 0px 5px 50px 0px #0169df, 0px 0px 0px 7px #5191da;
 }
 @media screen and (max-width: 576px) {
@@ -243,9 +243,12 @@ a:hover {
     box-shadow: 0px 7px 30px rgba(223, 45, 70, 0.75);
 }
 .profile-card__button.button--gray {
-    box-shadow: none;
-    background: #dcdcdc;
-    color: #142029;
+    background: linear-gradient(45deg, #bebebe, #c4c4c4);
+    box-shadow: 0px 4px 30px rgba(94, 94, 94, 0.35);
+    color: #1f2a33;
+}
+.profile-card__button.button--gray:hover {
+    box-shadow: 0px 7px 30px rgba(87, 87, 87, 0.75);
 }
 .profile-card-message {
     width: 100%;
@@ -259,6 +262,8 @@ a:hover {
     pointer-events: none;
     transition: all 0.3s;
 }
+
+/* traslladada parcialment a box */
 .profile-card-form {
     box-shadow: 0 4px 30px rgba(15, 22, 56, 0.35);
     max-width: 80%;
@@ -282,16 +287,6 @@ a:hover {
 @media screen and (max-width: 576px) {
     .profile-card-form {
         padding: 20px;
-    }
-}
-.profile-card-form__bottom {
-    justify-content: center;
-    display: flex;
-    margin-top: 10px;
-}
-@media screen and (max-width: 576px) {
-    .profile-card-form__bottom {
-        flex-wrap: wrap;
     }
 }
 
@@ -329,3 +324,5 @@ a:hover {
     border-radius: 12px;
     transition: all 0.3s;
 }
+
+/* -------------- */
