+++ frontend-app-skynuc/src/pages/home/home.module.css
@@ -28,7 +28,7 @@
     flex-direction: column;
 }
 
-h1{
+.__title{
     align-items: center;
     font-size: 65px;
     font-weight: bolder;
@@ -59,7 +59,7 @@ h1{
     margin: 10px 20px;
 }
 
-footer{
+.__container__footer{
     display: flex;
     justify-content: center;
     align-items: center;
@@ -69,7 +69,7 @@ footer{
     background-color: rgba(99, 107, 111, 0.9);
 }
 
-input{
+.__container__input{
     background: rgba(99, 107, 111, 0.9);
     width: 350px;
     text-align: center;
@@ -80,7 +80,7 @@ input{
     outline: none;
 }
 
-input::placeholder{
+.__container__input::placeholder{
     color: white;
     font-family: -apple-system, serif;
     outline: none;
